moved themes to external file
