+++ client/src/App.js
@@ -3,6 +3,7 @@ import GlobalStyle from './Services/GlobalStyle';
 import styled, {
   ThemeProvider,
 } from 'styled-components/macro';
+import themes from './Services/themes';
 import { Switch, Route } from 'react-router-dom';
 import { useEffect, useState } from 'react';
 import fetchData from './Services/fetchData';
@@ -129,30 +130,6 @@ function App() {
 export default App;
 
 //Set up for light and dark themes
-const LightTheme = {
-  fontColor: '#0B3142',
-  colorBackground: '#F5EDF7',
-  colorDarkButton: '#E9C1F0',
-  colorVeryLightButton: '#FFFFFF',
-  colorNavigation: '#EFDCF2',
-  cardBackgroundColor: 'rgba(239, 220, 242, 0.7)',
-};
-
-const DarkTheme = {
-  fontColor: '#fff',
-  colorBackground: '#2C2972',
-  colorDarkButton: '#6C24B4',
-  colorLighterButton: '#a626b7',
-  colorLightButton: '#b12dc3',
-  colorVeryLightButton: '#dc45f0',
-  colorNavigation: '#32A19E',
-  cardBackgroundColor: 'rgba(255, 255, 255, 0.1)',
-};
-
-const themes = {
-  light: LightTheme,
-  dark: DarkTheme,
-};
 
 const AppWrapper = styled.section`
   width: 100%;
