+++ client/src/Services/themes.js
@@ -1,0 +1,26 @@
+const LightTheme = {
+  fontColor: '#0B3142',
+  colorBackground: '#F5EDF7',
+  colorDarkButton: '#E9C1F0',
+  colorVeryLightButton: '#FFFFFF',
+  colorNavigation: '#EFDCF2',
+  cardBackgroundColor: 'rgba(239, 220, 242, 0.7)',
+};
+
+const DarkTheme = {
+  fontColor: '#fff',
+  colorBackground: '#2C2972',
+  colorDarkButton: '#6C24B4',
+  colorLighterButton: '#a626b7',
+  colorLightButton: '#b12dc3',
+  colorVeryLightButton: '#dc45f0',
+  colorNavigation: '#32A19E',
+  cardBackgroundColor: 'rgba(255, 255, 255, 0.1)',
+};
+
+const themes = {
+  light: LightTheme,
+  dark: DarkTheme,
+};
+
+export default themes;
